Bug 1290334 - Use single quotes for core_abspath. r=wtc
Double quotes allow all sorts of escaping, which are not necessarily
wanted. Moreover, it forces make to run a subshell instead of executing
the command directly on its own (because of the possibility of there
being something escaped between the double quotes), and that causes
problems when running under MINGW make with a MSYS shell.

When using single quotes, those problems go away, because make runs the
subprocess without an intermediate shell, and avoids a painful
MINGW->MSYS->MINGW transition.